// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface Product {
    id: bigint;
    displayOrder: bigint;
    name: string;
    createdAt: Time;
    stripeProductId: string;
    shape: string;
    viewCount: bigint;
    stripeProductDescription: string;
    category: string;
    inventoryCount: bigint;
    price: bigint;
    images: Array<ExternalBlob>;
}
export interface DescriptionTemplate {
    id: bigint;
    content: string;
    name: string;
    createdAt: Time;
}
export interface TransformationOutput {
    status: bigint;
    body: Uint8Array;
    headers: Array<http_header>;
}
export type Time = bigint;